Create core loader.js
